zabbix host group

+++ inventory-ui/inventory/src/features/zabbixQuery/zabbixActionQuery.ts
@@ -1,12 +1,17 @@
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
 import { zabbixActionsAPI } from '../../service/zabbixApi/zabbixActionAPI';
+import { useProjectId } from './projectIdHook';
 
 // Create Action
 export const useCreateZabbixAction = () => {
+  const projectId = useProjectId();
+  if (!projectId) {
+    throw new Error('Project ID is required to fetch host groups');
+  }
   const queryClient = useQueryClient();
   return useMutation({
-    mutationFn: ({ projectId, data }: { projectId: number; data: object }) =>
-      zabbixActionsAPI.create(projectId, data),
+    mutationFn: ({ params }: { params: object }) =>
+      zabbixActionsAPI.create(projectId, params),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['ZabbixActions'] });
     }
@@ -15,9 +20,13 @@ export const useCreateZabbixAction = () => {
 
 // Delete Action
 export const useDeleteZabbixAction = () => {
+  const projectId = useProjectId();
+  if (!projectId) {
+    throw new Error('Project ID is required to fetch host groups');
+  }
   const queryClient = useQueryClient();
   return useMutation({
-    mutationFn: ({ projectId, actionIds }: { projectId: number; actionIds: any }) =>
+    mutationFn: ({ actionIds }: { actionIds: string }) =>
       zabbixActionsAPI.delete(projectId, actionIds),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['ZabbixActions'] });
@@ -26,7 +35,11 @@ export const useDeleteZabbixAction = () => {
 };
 
 // Get Actions
-export const useZabbixActions = (projectId: number, params: object) => {
+export const useZabbixActions = (params: object) => {
+  const projectId = useProjectId();
+  if (!projectId) {
+    throw new Error('Project ID is required to fetch host groups');
+  }
   return useQuery({
     queryKey: ['ZabbixActions', projectId, params],
     queryFn: () => zabbixActionsAPI.get(projectId, params),
@@ -36,10 +49,14 @@ export const useZabbixActions = (projectId: number, params: object) => {
 
 // Update Action
 export const useUpdateZabbixAction = () => {
+  const projectId = useProjectId();
+  if (!projectId) {
+    throw new Error('Project ID is required to fetch host groups');
+  }
   const queryClient = useQueryClient();
   return useMutation({
-    mutationFn: ({ projectId, data }: { projectId: number; data: object }) =>
-      zabbixActionsAPI.update(projectId, data),
+    mutationFn: ({ params }: { params: object }) =>
+      zabbixActionsAPI.update(projectId, params),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['ZabbixActions'] });
     }
